Messed up by not fixing all the filepaths due to moving files/folders.
Everything is now fixed (hopefully).
