+++ src/pages/CalendarPage/components/EventList/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import PropTypes from 'prop-types';
 import moment from 'moment';
-import Table from '../Table';
+import Table from '../../../../components/Table';
 
 const eventDateComparator = (dateTimeA, dateTimeB) => {
   return moment(dateTimeA).diff(moment(dateTimeB));
