+++ src/pages/SignUpPage/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { Avatar, Card } from '@material-ui/core';
 import { withStyles } from '@material-ui/core/styles';
 
-import SignUpForm from '../../components/SignUpForm';
+import SignUpForm from './components/SignUpForm';
 
 import { createUserAccountFromSignup } from '../../services/auth';
 import HKN_TRIDENT_LOGO from '../../images/hkn-trident.png';
